analysis/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
pandas==2.2.1
openpyxl==3.1.2
httpx==0.27.0
python-dotenv==1.0.1
google-cloud-vision==3.7.0
anthropic==0.19.1
google-auth==2.28.1
beautifulsoup4==4.12.3
requests==2.31.0
numpy==1.26.4
scikit-learn==1.4.1
pytest==8.1.1
pytest-asyncio==0.23.5
